Rename old object library to model library
